Container for multiple problem templates

Container for multiple problem templates

Usage

build_problemset(template_list, prob = NULL, title = "", label = title)

Arguments

template_list

list of problem templates.

prob

optional vector of weights for use when sampling problems from the problem set. Defaults to NULL, representing a uniform distribution.

title

optional string used as a short title description of the whole set.

label

optional shorts label for the problem set. Defaults to the same value as title.

Value

a problemset object.

Examples

percent_problems = build_problemset(list(
  simple_percent,
  simple_percent_successes
), title="Percents")