Simple proportion/percent problem templates.

Collection of simple example problems involving proportions and/or percents.

Usage

simple_prop(seed = NULL)

pet_prop(seed = NULL)

simple_percent(seed = NULL)

simple_prop_percent(seed = NULL)

simple_successes(seed = NULL)

simple_percent_successes(seed = NULL)

dotplot_visual_prop(seed = NULL)

Arguments

seed

(optional) integer; pseudorandom seed used with withr::local_seed() inside the problem.

Value

compiled problem

Examples

simple_prop(seed=1)
#> $setup
#> $setup$total
#> [1] 26
#> 
#> $setup$successes
#> [1] 1
#> 
#> $setup$failures
#> [1] 25
#> 
#> $setup$proportion
#> [1] 0.038
#> 
#> 
#> $question
#> In a survey of 26 people, 1 people said "yes" to a particular question. What proportion of the people said "yes"?
#> 
#> $answer
#> [1] 0.038
#> 
#> $solution
#> The proportion is $\frac{1}{26} = 0.038$.
#> 
pet_prop(seed=2)
#> $setup
#> $setup$total
#> [1] 25
#> 
#> $setup$successes
#> [1] 6
#> 
#> $setup$failures
#> [1] 19
#> 
#> $setup$proportion
#> [1] 0.24
#> 
#> 
#> $question
#> In a survey of 25 people, 6 people said they had a pet. What proportion of the people had a pet?
#> 
#> $answer
#> [1] 0.24
#> 
#> $solution
#> The proportion of people with a pet is $\frac{6}{25} = 0.24$.
#> 
simple_percent(seed=3)
#> $setup
#> $setup$total
#> [1] 24
#> 
#> $setup$successes
#> [1] 7
#> 
#> $setup$failures
#> [1] 17
#> 
#> $setup$proportion
#> [1] 0.292
#> 
#> 
#> $question
#> In a survey of 24 people, 7 people said "yes" to a particular question. What percent of the people said "yes"?
#> 
#> $answer
#> [1] 29.2
#> 
#> $solution
#> The percentage is $\frac{7}{24}\cdot 100\% = 29.2\%$.
#> 
simple_prop_percent(seed=4)
#> $setup
#> $setup$total
#> [1] 31
#> 
#> $setup$successes
#> [1] 19
#> 
#> $setup$failures
#> [1] 12
#> 
#> $setup$proportion
#> [1] 0.613
#> 
#> 
#> $question
#> In a survey of 31 people, 19 people said "yes" to a particular question.
#> 
#> $subquestions
#> [1] "What proportion of the people said \"yes\"?"
#> [2] "What percent of the people said \"yes\"?"   
#> 
#> $figures
#> NULL
#> 
#> $tables
#> NULL
#> 
#> $solution
#> 
#> $subsolutions
#> [1] "The proportion is $\\frac{19}{31} = 0.613$."               
#> [2] "The percentage is $\\frac{19}{31}\\cdot 100\\% = 61.3\\%$."
#> 
#> $sol_figures
#> NULL
#> 
#> $sol_tables
#> NULL
#> 
simple_successes(seed=5)
#> $setup
#> $setup$total
#> [1] 25
#> 
#> $setup$successes
#> [1] 15
#> 
#> $setup$failures
#> [1] 10
#> 
#> $setup$proportion
#> [1] 0.6
#> 
#> 
#> $question
#> In a survey of 25 people, the proportion of the people who said "yes" to a particular question was 0.6. How many of the people said "yes"?
#> 
#> $answer
#> [1] 15
#> 
#> $solution
#> The number of people who said "yes" is $25 \cdot 0.6 = 15$.
#> 
simple_percent_successes(seed=6)
#> $setup
#> $setup$total
#> [1] 31
#> 
#> $setup$successes
#> [1] 13
#> 
#> $setup$failures
#> [1] 18
#> 
#> $setup$proportion
#> [1] 0.419
#> 
#> 
#> $question
#> In a survey of 31 people, 41.9% of the people said "yes" to a particular question. How many of the people said "yes"?
#> 
#> $answer
#> [1] 13
#> 
#> $solution
#> The number of people who said "yes" is $31 \cdot 0.419 = 13$.
#> 
dotplot_visual_prop(seed=7)
#> $setup
#> $setup$n_samps
#> [1] 23
#> 
#> $setup$mu
#> [1] 20
#> 
#> $setup$s
#> [1] 2
#> 
#> $setup$samp
#>  [1] 17.60646 18.61141 19.17541 18.05865 18.10544 21.49628 19.76609 20.30532
#>  [9] 24.37996 20.71397 25.43350 24.56290 20.64804 23.79213 20.93536 18.21240
#> [17] 19.38534 19.99036 21.97633 21.67950 21.41068 22.61193 17.22401
#> 
#> $setup$breaks
#>  [1] 16.5 17.0 17.5 18.0 18.5 19.0 19.5 20.0 20.5 21.0 21.5 22.0 22.5 23.0 23.5
#> [16] 24.0 24.5 25.0 25.5 26.0
#> 
#> $setup$cutoff
#> [1] 17.5
#> 
#> $setup$gtlt
#> [1] ">"
#> 
#> $setup$successes
#> [1] 22
#> 
#> $setup$proportion
#> [1] 0.957
#> 
#> 
#> $question
#> What proportion of the 23 points shown are greater than 17.5 (dashed line)?
#> 
#> $answer
#> [1] 0.957
#> 
#> $solution
#> We see that 22 of the points (colored red here) are greater than 17.5, so the proportion is $\frac{22}{23} = 0.957$.
#> 
#> $figures
#> $figures[[1]]
#> function () 
#> {
#>     func(vars)
#> }
#> <bytecode: 0x560ef7c9a168>
#> <environment: 0x560ef71c1f80>
#> 
#> 
#> $sol_figures
#> $sol_figures[[1]]
#> function () 
#> {
#>     func(vars)
#> }
#> <bytecode: 0x560ef7c9a168>
#> <environment: 0x560ef71b9bb8>
#> 
#>